[Bugs] Multiple bugs:
* Electro ray after first shoot starts to flicker
* Electro ray shoots at first in the center, like there is some target
* When new objects are spawned, like spaceships or loot, they appear for a short moment of time in the center of the scene.

+++ PHEngine/PHEngine/src/Core/GameCore/Physics/PhysicsDescriptors/PhysicsDescriptor.h
@@ -86,7 +86,7 @@ public:
 
     const std::shared_ptr<CollisionShapeBase>& GetShape() const;
 
-    size_t GetId() const;
+    int32_t GetId() const;
 
     float GetMass() const;
